Fix metrics reporting
Previously, filtering by metric name with an empty filter would return
empty metrics.

With an upgrade of prometheus simpleclient, this is no longer the
case, and simpleclient just returns an empty set of metrics.

This appears to be because simpleclient doesn't properly register the
metrics in CollectorRegistry when CollectorRegistry.register is
called. This did not make a difference previously because when calling
the metric filtering methods with an empty filter would return all
collectors, but now that logic seems to have changed in simpleclient
and this is no longer the case.

Since we do not use filtering anyway, we remove the support for it in
PrometheusMetrics, which now returns all metrics as previously.