Add implicit method toErrorRepr to errors
So they can be easily converted to ErrorRepresentation outside libats.

We do this instead of providing `io.circe.Encoder[Error]` as we want
to convert all errors to a common format, `ErrorRepresentation`, and
providing an `Encoder` to encode `Error => ErrorRepresentation =>
Json` would work but would be likely to cause confusion.