Refactor kafka client to support a `Source[CommitableMessage[T]]`
To allow libats users more flexibility when building up streams of
messages, for example, batching messages to optimize processing.

This allows users more fine grained control over the behavior of the
message stream, but the trade off is that the lifecycle and monitoring
of the stream as well as kafka offset commiting and metrics, need to
be implemented explicitly.